Changes
1. Bug in renaming solved. If editor was not having any data, then
   renaming it was deleting old file.
2. Font family change.
3. dynamic css change.

+++ script.js
@@ -123,7 +123,9 @@ editor_div.addEventListener('focusout', () => {
 fileName.addEventListener('focusout', () => {
     let fname = (fileName.value).trim()
     renameTab(fname);
-    if (!fileExist(currentFileName) && fname != currentFileName) {
+    let editor = document.querySelector('.active').innerHTML;
+    if (!fileExist(currentFileName) && fname != currentFileName && editor != '') {
+        console.log('here--1', currentFileName)
         localStorage.setItem(fname, getEditorData(currentFileName));
         localStorage.removeItem(currentFileName);
         currentFileName = fname;
@@ -201,7 +203,7 @@ function changeWidth () {
     } else if (fileName.value.length > 46) {
         fileName.style.width = 600 + 'px';
     } else {
-        fileName.style.width = (fileName.value.length - 1) + 'ch';
+        fileName.style.width = (fileName.value.length + 1) + 'ch';
     }
 }
 // #endregion
